Implement Email Invitation Management and Enhance Dealership Invitation Flow
- Introduced new use cases for accepting and rejecting dealership invitations, including corresponding DTOs and controllers.
- Updated the Dealership module to handle invitation management, ensuring proper validation and error handling.
- Enhanced email services to support sending invitations and notifications for accepted or rejected invitations via BullMQ.
- Refactored existing dealership invitation logic to improve clarity and maintainability, including role validation for dealership members.
- Updated Postman collection to include new endpoints for dealership invitation acceptance and rejection, improving API documentation.

+++ src/contexts/auth/services/auth.service.ts
@@ -63,6 +63,10 @@ export class AuthService {
   }
 
   async createSession(user: User, request: Request): Promise<{ session_id: string, refreshToken_hash: string }> {
+    const currentSession = await this.sessionService.findOneByIpAddress(request.ip);
+    if (currentSession) {
+      await this.sessionService.delete(currentSession.id);
+    }
     const session = await this.sessionService.create(user, request);
     const refreshToken = await this.refreshTokenService.createForSession(user, session);
 
@@ -104,6 +108,9 @@ export class AuthService {
 
   async logout(request: Request) {
     const session = await this.sessionService.findOneByIpAddress(request.ip);
+    if (!session) {
+      throw new UnauthorizedException("Sesión no encontrada");
+    }
     await this.sessionService.delete(session.id);
   }
